+++ R/CMap.R
@@ -111,9 +111,9 @@ prepareCMapZscores <- function(file, zscoresID=NULL) {
 #' @export
 #'
 #' @examples
-#' \donttest{
 #' metadata <- loadCMapData("cmapMetadata.txt", "metadata")
 #' metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
+#' \dontrun{
 #' perts <- prepareCMapPerturbations(metadata, "cmapZscores.gctx",
 #'                                   "cmapGeneInfo.txt")
 #' zscores <- loadCMapZscores(perts[ , 1:10])
@@ -314,9 +314,7 @@ getCMapConditions <- function(metadata, cellLine=NULL, timepoint=NULL,
 #' @export
 #'
 #' @examples
-#' \dontrun{
 #' cmapMetadata <- loadCMapData("cmapMetadata.txt", "metadata")
-#' }
 #' filterCMapMetadata(cmapMetadata, cellLine="HEPG2", timepoint="2 h",
 #'                    dosage="25 ng/mL")
 filterCMapMetadata <- function(metadata, cellLine=NULL, timepoint=NULL,
@@ -373,9 +371,9 @@ filterCMapMetadata <- function(metadata, cellLine=NULL, timepoint=NULL,
 #' @export
 #'
 #' @examples
-#' \dontrun{
 #' metadata <- loadCMapData("cmapMetadata.txt", "metadata")
 #' metadata <- filterCMapMetadata(metadata, cellLine="HepG2")
+#' \dontrun{
 #' prepareCMapPerturbations(metadata, "cmapZscores.gctx", "cmapGeneInfo.txt")
 #' }
 prepareCMapPerturbations <- function(metadata, zscores, geneInfo,